Add files via upload
To see the whole list of documents written by me, we can write db.items.find(). The data stored in the data directory will never get deleted until you do it manually. The process takes up the commands one by one and stores it in the data.

If we paste the objects created by me in the powershell and run the command db.items.find(), we get the output as follows: